+++ src/Sulu/Bundle/ContentBundle/Form/Type/BasePageDocumentType.php
@@ -14,6 +14,7 @@ namespace Sulu\Bundle\ContentBundle\Form\Type;
 use Sulu\Component\Content\Form\Type\DocumentObjectType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\CollectionType;
+use Symfony\Component\Form\Extension\Core\Type\DateType;
 use Symfony\Component\Form\Extension\Core\Type\IntegerType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
@@ -59,6 +60,12 @@ abstract class BasePageDocumentType extends AbstractStructureBehaviorType
         $builder->add('workflowStage', IntegerType::class);
         $builder->add('shadowLocaleEnabled', CheckboxType::class);
         $builder->add('shadowLocale', TextType::class); // TODO: Should be choice of available shadow locales
+        $builder->add(
+            'authored',
+            DateType::class,
+            ['widget' => 'single_text', 'model_timezone' => 'UTC', 'view_timezone' => 'UTC']
+        );
+        $builder->add('author', TextType::class);
         $builder->setAttribute('webspace_key', $options['webspace_key']);
     }
 }