+++ src/Sulu/Bundle/WebsiteBundle/Resolver/StructureResolver.php
@@ -14,6 +14,7 @@ namespace Sulu\Bundle\WebsiteBundle\Resolver;
 use Sulu\Component\Content\Compat\Structure\PageBridge;
 use Sulu\Component\Content\Compat\StructureInterface;
 use Sulu\Component\Content\ContentTypeManagerInterface;
+use Sulu\Component\Content\Document\Behavior\LocalizedAuthorBehavior;
 use Sulu\Component\Content\Extension\ExtensionManagerInterface;
 
 /**
@@ -69,6 +70,12 @@ class StructureResolver implements StructureResolverInterface
             foreach ($data['extension'] as $name => $value) {
                 $extension = $this->extensionManager->getExtension($structure->getKey(), $name);
                 $data['extension'][$name] = $extension->getContentData($value);
+            }
+
+            $document = $structure->getDocument();
+            if ($document instanceof LocalizedAuthorBehavior) {
+                $data['authored'] = $document->getAuthored();
+                $data['author'] = $document->getAuthor();
             }
         }