+++ src/Sulu/Component/Content/Document/Behavior/AuthorBehavior.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * This file is part of Sulu.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Component\Content\Document\Behavior;
+
+/**
+ * Adds the user which is defined as author and an authored date (without locale).
+ */
+interface AuthorBehavior extends LocalizedAuthorBehavior
+{
+}